README tweaks
A few tweaks to formatting.

Re: installing `npm` globally, just no. That's typically only done for CLI tools that are part of build processes. E.g. `yarn`, maybe `jest`... that sort of thing.

Note: For something like this, I would consider providing the examples as vanilla JS.  It makes them a bit easier to read, and is more approachable for non-TS users.